ajout test fibo 7, test fibo 10, test pap (sub 10 4 en pap)

+++ src/interpreter/primitives.rs
@@ -32,7 +32,7 @@ pub fn is_primitive(nom: &String) -> bool {
     PRIMITIVES.contains(&nom.as_str())
 }
 
-pub fn has_args(nom : &String, length:usize) -> usize {
+pub fn has_args(nom : &String, length:i32) -> i32 {
     if nom.clone().eq("not") {
         return length - 1;
     }
